Added JWT Security

+++ src/main/java/com/calendar/services/EventServiceImpl.java
@@ -2,12 +2,12 @@ package com.calendar.services;
 
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import com.calendar.entities.Event;
 import com.calendar.entities.User;
 import com.calendar.exceptions.EventNotFoundException;
-import com.calendar.exceptions.UserNotFoundException;
 import com.calendar.payloads.EventDTO;
 import com.calendar.repositories.EventRepo;
 import com.calendar.repositories.UserRepo;
@@ -23,11 +23,15 @@ public class EventServiceImpl implements EventService {
 
 	@Autowired
 	private ModelMapper modelMapper;
-
+	
 	@Override
-	public EventDTO createEvent(EventDTO eventDTO, String email) throws UserNotFoundException {
-		User user = userRepo.findById(email)
-				.orElseThrow(() -> new UserNotFoundException("User not found with email: " + email));
+	public EventDTO createEvent(EventDTO eventDTO) {
+		String email = (String) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+		
+//		User user = userRepo.findById(email)
+//				.orElseThrow(() -> new UserNotFoundException("User not found with email: " + email));
+		
+		User user = userRepo.findById(email).get();
 
 		Event event = modelMapper.map(eventDTO, Event.class);
 
@@ -51,6 +55,12 @@ public class EventServiceImpl implements EventService {
 	public EventDTO updateEvent(Long eventId, EventDTO eventDTO) throws EventNotFoundException {
 		Event event = eventRepo.findById(eventId)
 				.orElseThrow(() -> new EventNotFoundException("Event not found with event id: " + eventId));
+		
+		String email = (String) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+		
+		if(!email.equals(event.getUser().getEmail())) {
+			throw new EventNotFoundException("You have not scheduled any event with event id " + eventId + " to update !!");
+		}
 
 		event.setStartDate(eventDTO.getStartDate());
 		event.setEndDate(eventDTO.getEndDate());
@@ -72,13 +82,19 @@ public class EventServiceImpl implements EventService {
 	}
 
 	@Override
-	public EventDTO deleteEvent(Long eventId) throws EventNotFoundException {
+	public String deleteEvent(Long eventId) throws EventNotFoundException {
 		Event event = eventRepo.findById(eventId)
 				.orElseThrow(() -> new EventNotFoundException("Event not found with event id: " + eventId));
 		
+		String email = (String) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+		
+		if(!email.equals(event.getUser().getEmail())) {
+			throw new EventNotFoundException("You have not scheduled any event with id " + eventId + " to delete !!");
+		}
+		
 		eventRepo.delete(event);
 		
-		return modelMapper.map(event, EventDTO.class);
+		return "Event with id " + eventId + " deleted successfully !!";
 	}
 
 }
